+++ lib/src/components/parallax_component.dart
@@ -13,14 +13,17 @@ import 'position_component.dart';
 extension ParallaxComponentExtension on Game {
   Future<ParallaxComponent> loadParallaxComponent(
     List<String> paths, {
+    Vector2 size,
     Vector2 baseVelocity,
     Vector2 velocityMultiplierDelta,
     ImageRepeat repeat = ImageRepeat.repeatX,
     Alignment alignment = Alignment.bottomLeft,
     LayerFill fill = LayerFill.height,
-  }) {
-    return ParallaxComponent.load(
+  }) async {
+    final componentSize = size ?? this.size;
+    final component = await ParallaxComponent.load(
       paths,
+      size: componentSize,
       baseVelocity: baseVelocity,
       velocityMultiplierDelta: velocityMultiplierDelta,
       repeat: repeat,
@@ -28,41 +31,47 @@ extension ParallaxComponentExtension on Game {
       fill: fill,
       images: images,
     );
+
+    return component..size.setFrom(componentSize);
   }
 }
 
 /// A full parallax, several layers of images drawn out on the screen and each
 /// layer moves with different velocities to give an effect of depth.
 class ParallaxComponent extends PositionComponent {
-  final Parallax parallax;
+  Parallax _parallax;
 
-  ParallaxComponent(this.parallax);
+  Parallax get parallax => _parallax;
+  set parallax(Parallax p) {
+    _parallax = p;
+    _parallax.resize(size);
+  }
+
+  /// Creates a component with an empty parallax which can be set later
+  ParallaxComponent();
+
+  /// Creates a component from a [Parallax] object
+  ParallaxComponent.fromParallax(this._parallax);
 
   @mustCallSuper
   @override
   void onGameResize(Vector2 size) {
     super.onGameResize(size);
     this.size = size;
-    parallax.resize(size);
+    parallax?.resize(size);
   }
 
   @override
   void update(double t) {
     super.update(t);
-    parallax.update(t);
+    parallax?.update(t);
   }
 
   @mustCallSuper
   @override
   void render(Canvas canvas) {
     super.render(canvas);
-    canvas.save();
-    parallax.layers.forEach((layer) {
-      canvas.save();
-      layer.render(canvas);
-      canvas.restore();
-    });
-    canvas.restore();
+    parallax?.render(canvas);
   }
 
   /// Note that this method only should be used if all of your layers should
@@ -70,7 +79,7 @@ class ParallaxComponent extends PositionComponent {
   /// and filled), otherwise load the [ParallaxLayer]s individually and use the
   /// normal constructor.
   ///
-  /// [load] takes a list of paths to all the images that you want to use in the
+  /// [load] takes a list of paths to all the images and a size that you want to use in the
   /// parallax.
   /// Optionally arguments for the [baseVelocity] and [layerDelta] can be passed
   /// in, [baseVelocity] defines what the base velocity of the layers should be
@@ -84,6 +93,7 @@ class ParallaxComponent extends PositionComponent {
   /// If no image cache is set, the global flame cache is used.
   static Future<ParallaxComponent> load(
     List<String> paths, {
+    Vector2 size,
     Vector2 baseVelocity,
     Vector2 velocityMultiplierDelta,
     ImageRepeat repeat = ImageRepeat.repeatX,
@@ -91,9 +101,10 @@ class ParallaxComponent extends PositionComponent {
     LayerFill fill = LayerFill.height,
     Images images,
   }) async {
-    return ParallaxComponent(
+    final component = ParallaxComponent.fromParallax(
       await Parallax.load(
         paths,
+        size,
         baseVelocity: baseVelocity,
         velocityMultiplierDelta: velocityMultiplierDelta,
         repeat: repeat,
@@ -102,5 +113,11 @@ class ParallaxComponent extends PositionComponent {
         images: images,
       ),
     );
+
+    if (size != null) {
+      component.size.setFrom(size);
+    }
+
+    return component;
   }
 }