+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## [next]
+ - Moving device related methods (like `fullScreen`) from `util.dart` to `device.dart`
+ - Moving render functions from `util.dart` to `extensions/canvas.dart`
+ - Adapting ParallaxComponent contructors to match the pattern followed on other components
+ - Adapting SpriteBatchComponent constructors to match the pattern used on other components
+ - Improving Parallax APIs regarding handling its size and the use outside FCS
+ - Enabling direct import of Sprite and SpriteAnimation
+ - Renamed `Composition` to `ImageComposition` to prevent confusion with the composition component
+ - Added `rotation` and `anchor` arguments to `ImageComposition.add`
+ - Added `Image` extensions
+ - Added `Color` extensions
+ - Change RaisedButton to ElevatedButton in timer example
+ - Overhaul the draggables api to fix issues relating to local vs global positions
+ - Preventing errors caused by the premature use of size property on game
+ - Added a hitbox mixin for PositionComponent to make more accurate gestures
+ - Added a collision detection system
+ - Added geometrical shapes
+ - Fix `SpriteAnimationComponent.shouldRemove` use `Component.shouldRemove`
+ - Add assertion to make sure Draggables are safe to add
+ - Add utility methods to the Anchor class to make it more "enum like"
+
 ## 1.0.0-rc6
  - Use `Offset` type directly in `JoystickAction.update` calculations
  - Changed `parseAnchor` in `examples/widgets` to throw an exception instead of returning null when it cannot parse an anchor name
@@ -24,7 +45,6 @@
  - Move files to comply with the dart package layout convention
  - Fix gesture detection bug of children of `PositionComponent`
  - The `game` argument on `GameWidget` is now required
- - Add hitbox mixin for PositionComponent to make more accurate gestures
 
 ## 1.0.0-rc5
  - Option for overlays to be already visible on the GameWidget
@@ -100,6 +120,13 @@
  - Generalize effects api to include all components
  - Extract all the audio related capabilities to a new package, flame_audio
  - Fix bug that sprite crashes without a position
+
+## 0.29.1-beta
+ - Fixing Game Render Box for flutter >= 1.25
+
+## 0.29.0
+- Update audioplayers to latest version (now `assets` will not be added to prefixes automatically)
+- Fix lint issues with 0.28.0
 
 ## 0.28.0
 - Fix spriteAsWidget deprecation message